fix: changed from init to Long tho match types

+++ app/src/main/java/com/task/management/workflow/calendar/domain/CreateEventRequest.kt
@@ -3,8 +3,8 @@ package com.task.management.workflow.calendar.domain
 import java.time.LocalDate
 
 class CreateEventRequest(
-    val projectId: Int,
-    val userId: Int,
+    val projectId: Long,
+    val userId: Long,
     val title: String,
     val description: String,
     var dueDate: String
